fix(ci): CI lint/repo-shape drift on consumption-ui branch
Ruff format drift on three server files, a stale generated.ts line-count
allowlist entry, and two desktop components pushed over the 500-line
component cap by the new consumption card / limits editor UI. Split
SidebarAccountFooter and OrganizationBudgetsPane into their own
sub-components rather than allowlisting them.
