+++ scripts/max_lines_allowlist.txt
@@ -24,7 +24,7 @@ anyharness/crates/anyharness-lib/src/domains/workspaces/retire_preflight.rs 613 
 anyharness/crates/anyharness-lib/src/live/sessions/probe.rs 654 catalog probe with per-harness model-source fallbacks and tests
 anyharness/crates/anyharness-lib/src/persistence/custom_migrations.rs 1112 custom migration with legacy schema coverage
 anyharness/sdk-react/src/hooks/sessions.ts 774 existing SDK React oversized file
-cloud/sdk/src/types/generated.ts 761 existing generated cloud SDK type surface
+cloud/sdk/src/types/generated.ts 773 existing generated cloud SDK type surface (+billing budget-limit/usage types)
 anyharness/sdk/src/reducer/__tests__/transcript.test.ts 1919 existing SDK oversized test file
 anyharness/sdk/src/reducer/transcript.ts 1591 existing SDK oversized file
 apps/desktop/src-tauri/src/commands/ssh_tunnel.rs 729 existing desktop native oversized file
